+++ lib/services/saml_idp_service_provider.go
@@ -20,11 +20,13 @@ package services
 
 import (
 	"context"
+	"fmt"
 	"net/url"
 	"slices"
 	"strings"
 
 	"github.com/crewjam/saml"
+	"github.com/crewjam/saml/samlsp"
 	"github.com/gravitational/trace"
 	log "github.com/sirupsen/logrus"
 
@@ -151,8 +153,8 @@ func FilterSAMLEntityDescriptor(ed *saml.EntityDescriptor, quiet bool) error {
 		ed.SPSSODescriptors[i].AssertionConsumerServices = filtered
 	}
 
-	if filteredCount == 0 && originalCount != 0 {
-		return trace.BadParameter("no AssertionConsumerService bindings for entity %q passed validation", ed.EntityID)
+	if filteredCount != originalCount {
+		return trace.BadParameter("Entity descriptor for entity %q contains unsupported AssertionConsumerService binding or location", ed.EntityID)
 	}
 
 	return nil
@@ -162,6 +164,43 @@ func FilterSAMLEntityDescriptor(ed *saml.EntityDescriptor, quiet bool) error {
 // commonly used in xss payload. This is not a comprehensive list but is only
 // meant to increase the ost of xss payload.
 const invalidSAMLIdPACSURLChars = `<>"!;`
+
+// SAMLACSInputFilteringThreshold defines level of strictness for entity descriptor filtering.
+type SAMLACSInputFilteringThreshold string
+
+const (
+	// SAMLACSInputStrictFilter indicates ValidateAndFilterEntityDescriptor to return an error on
+	// any instance of unsupported ACS value.
+	SAMLACSInputStrictFilter SAMLACSInputFilteringThreshold = "SAMLACSInputStrictFilter"
+	// SAMLACSInputStrictFilter indicates ValidateAndFilterEntityDescriptor to ignore an error on
+	// any instance of unsupported ACS value.
+	SAMLACSInputPermissiveFilter SAMLACSInputFilteringThreshold = "SAMLACSInputPermissiveFilter"
+)
+
+// ValidateAndFilterEntityDescriptor validates entity id and ACS value. It specifically does:
+//   - checks for a valid entity descriptor XML format.
+//   - checks for a matching entity ID field in both the entity_id field and entity ID contained in the value of
+//     entity_descriptor field.
+//   - performs filtering on the Assertion Consumer service (ACS) binding format or its location URL endpoint.
+//     filterThreshold dictates if ValidateAndFilterEntityDescriptor should return or ignore error on filtering result.
+func ValidateAndFilterEntityDescriptor(sp types.SAMLIdPServiceProvider, filterThreshold SAMLACSInputFilteringThreshold) error {
+	edOriginal, err := samlsp.ParseMetadata([]byte(sp.GetEntityDescriptor()))
+	if err != nil {
+		return trace.BadParameter("invalid entity descriptor for SAML IdP Service Provider %q: %v", sp.GetEntityID(), err)
+	}
+
+	if edOriginal.EntityID != sp.GetEntityID() {
+		return trace.BadParameter("entity ID parsed from the entity descriptor does not match the entity ID in the SAML IdP service provider object")
+	}
+
+	if err := FilterSAMLEntityDescriptor(edOriginal, false /* quiet */); err != nil {
+		if filterThreshold == SAMLACSInputStrictFilter {
+			return trace.BadParameter("Entity descriptor for SAML IdP Service Provider %q contains unsupported ACS bindings: %v", sp.GetEntityID(), err)
+		}
+	}
+
+	return nil
+}
 
 // validateAssertionConsumerServicesEndpoint ensures that the Assertion Consumer Service location
 // is a valid HTTPS endpoint.
@@ -195,3 +234,19 @@ func ValidateSAMLIdPACSURLAndRelayStateInputs(sp types.SAMLIdPServiceProvider) e
 
 	return nil
 }
+
+// NewSAMLTestSPMetadata creates a new entity descriptor for tests.
+func NewSAMLTestSPMetadata(entityID, acsURL string) string {
+	return fmt.Sprintf(samlTestSPMetadata, entityID, acsURL)
+}
+
+// samlTestSPMetadata mimics metadata format generated by saml.ServiceProvider.Metadata()
+// Metadata already has one valid ACS endpoint.
+const samlTestSPMetadata = `<EntityDescriptor xmlns="urn:oasis:names:tc:SAML:2.0:metadata" validUntil="2023-12-09T23:43:58.16Z" entityID="%s">
+<SPSSODescriptor xmlns="urn:oasis:names:tc:SAML:2.0:metadata" validUntil="2023-12-09T23:43:58.16Z" protocolSupportEnumeration="urn:oasis:names:tc:SAML:2.0:protocol" AuthnRequestsSigned="false" WantAssertionsSigned="true">
+  <NameIDFormat>urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified</NameIDFormat>
+  <AssertionConsumerService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Location="%s" index="1"></AssertionConsumerService>
+  <AssertionConsumerService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Location="https://example.com/acs" index="1"></AssertionConsumerService>
+</SPSSODescriptor>
+</EntityDescriptor>
+`
